moved nav bar to navigation
