+++ src/components/Header/Header.css
@@ -20,46 +20,4 @@
   color: white;
 }
 
-.header__navbar {
-  display: flex;
-  gap: 42px;
-  padding-right: 104px;
-}
 
-.header__button {
-  position: relative;
-  display: flex;
-  align-items: center;
-  border: none;
-  background-color: transparent;
-  font-family: Roboto, "Open Sans", "Helvetica Neue", sans-serif;
-  font-weight: 500;
-  font-size: 16px;
-  line-height: 24px;
-  cursor: pointer;
-  padding: 0;
-  color: white;
-  text-decoration: none;
-}
-
-.header__button--active::after {
-  content: "";
-  position: absolute;
-  left: 50%;
-  transform: translateX(-50%);
-  bottom: -26px;
-  width: 100%;
-  height: 3px;
-  background: #fff;
-  border-radius: 2px;
-}
-
-.header__button-signin {
-  display: flex;
-  justify-content: center;
-  gap: 15px;
-  border: solid 1px;
-  border-radius: 100px;
-  padding: 12px;
-  width: 152px;
-}
