+++ src/components/Navigation/Navigation.css
@@ -1,0 +1,85 @@
+.nav__button {
+  background-color: #4caf50;
+  color: white;
+  padding: 16px;
+  font-size: 16px;
+  border: none;
+  cursor: pointer;
+}
+
+.nav {
+  position: relative;
+  display: inline-block;
+}
+
+.nav__content {
+  display: none;
+  position: absolute;
+  background-color: #f9f9f9;
+  min-width: 160px;
+  box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
+  z-index: 1;
+}
+
+.nav__content a {
+  color: black;
+  padding: 12px 16px;
+  text-decoration: none;
+  display: block;
+}
+
+.nav__content a:hover {
+  background-color: #f1f1f1;
+}
+
+.nav:hover .dropdown-content {
+  display: block;
+}
+
+.nav:hover .dropbtn {
+  background-color: #3e8e41;
+}
+
+.nav__navbar {
+  display: flex;
+  gap: 42px;
+  padding-right: 104px;
+}
+
+.nav__button {
+  position: relative;
+  display: flex;
+  align-items: center;
+  border: none;
+  background-color: transparent;
+  font-family: Roboto, "Open Sans", "Helvetica Neue", sans-serif;
+  font-weight: 500;
+  font-size: 16px;
+  line-height: 24px;
+  cursor: pointer;
+  padding: 0;
+  color: white;
+  text-decoration: none;
+}
+
+.nav__button--active::after {
+  content: "";
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+  bottom: -26px;
+  width: 100%;
+  height: 3px;
+  background: #fff;
+  border-radius: 2px;
+}
+
+.nav__button-signin {
+  display: flex;
+  justify-content: center;
+  gap: 15px;
+  border: solid 1px;
+  border-radius: 100px;
+  padding: 12px;
+  width: 152px;
+}
